+++ postgres-single/CHANGELOG.md
@@ -1,3 +1,4 @@
 0.0
 
 * Initial postgres-single image
+* Fix postgres setup